Refactored current code
