--- ProjectManagement/ProjectManagement.Core/UseCases/Projects/Queries/GetProjects/Dto/StepDto.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace ProjectManagement.Core.UseCases.Projects.Queries.GetProjects.Dto
-{
-    public class StepDto
-    {
-    }
-}
